Fix ancestor_id float dtype at source in alifestd_try_add_ancestor_id_col
When id column has float dtype (e.g., empty pandas DataFrames),
ancestor_id was being cast to float64, causing downstream numba
errors. Now detects float id dtype and uses int64 instead, since
id should always be integer.

Reverts the int64 cast band-aids in _alifestd_mark_csr_children_asexual.

+++ phyloframe/legacy/_alifestd_mark_csr_children_asexual.py
@@ -120,11 +120,9 @@ def alifestd_mark_csr_children_asexual(
         phylogeny_df = alifestd_topological_sort(phylogeny_df, mutate=True)
 
     if alifestd_has_contiguous_ids(phylogeny_df):
-        ancestor_ids = phylogeny_df["ancestor_id"].to_numpy().astype(np.int64)
+        ancestor_ids = phylogeny_df["ancestor_id"].to_numpy()
         if "csr_offsets" in phylogeny_df.columns:
-            csr_offsets = (
-                phylogeny_df["csr_offsets"].to_numpy().astype(np.int64)
-            )
+            csr_offsets = phylogeny_df["csr_offsets"].to_numpy()
         else:
             csr_offsets = _alifestd_mark_csr_offsets_asexual_fast_path(
                 ancestor_ids,
